Fixed the Menu, added new prefabs (forest, water, and ending), added background music and sound effects, added more controls and fixed some code.
Fixed the menu that it's now in its own scene and is able to load up the farmscene. Added 2 new levels and the ending. Added the background music and sound effects to make the game more enjoyable. Added more controls for a more smoother gameplay, and fixed the code for adding rocks and grass to make it work.